requirements.txt
certifi==2020.12.5
chardet==4.0.0
falcon==3.0.0
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
pygments==2.9.0
pymysql==1.0.2
pysocks==1.7.1
pyyaml==5.4.1
requests-toolbelt==0.9.1
requests[socks]==2.25.1
urllib3==1.26.4
